+++ src/main/java/com/expediagroup/dropwizard/bundle/configuration/freemarker/DefaultSystemPropertiesProvider.java
@@ -1,0 +1,10 @@
+package com.expediagroup.dropwizard.bundle.configuration.freemarker;
+
+import java.util.Properties;
+
+public class DefaultSystemPropertiesProvider implements SystemPropertiesProvider {
+    @Override
+    public Properties getSystemProperties() {
+        return System.getProperties();
+    }
+}