+++ src/main/java/com/expediagroup/dropwizard/bundle/configuration/freemarker/EnvironmentProvider.java
@@ -1,0 +1,7 @@
+package com.expediagroup.dropwizard.bundle.configuration.freemarker;
+
+import java.util.Map;
+
+public interface EnvironmentProvider {
+    Map<String, String> getEnvironment();
+}