+++ src/main/java/com/expediagroup/dropwizard/bundle/configuration/freemarker/SystemPropertiesProvider.java
@@ -1,0 +1,7 @@
+package com.expediagroup.dropwizard.bundle.configuration.freemarker;
+
+import java.util.Properties;
+
+public interface SystemPropertiesProvider {
+    Properties getSystemProperties();
+}