+++ src/main/java/de/thomaskrille/dropwizard_template_config/EnvironmentProvider.java
@@ -1,7 +1,0 @@
-package de.thomaskrille.dropwizard_template_config;
-
-import java.util.Map;
-
-public interface EnvironmentProvider {
-    Map<String, String> getEnvironment();
-}