+++ src/main/java/de/thomaskrille/dropwizard_template_config/SystemPropertiesProvider.java
@@ -1,7 +1,0 @@
-package de.thomaskrille.dropwizard_template_config;
-
-import java.util.Properties;
-
-public interface SystemPropertiesProvider {
-    Properties getSystemProperties();
-}