+++ src/test/java/com/expediagroup/dropwizard/bundle/configuration/freemarker/TestConfigSourceProvider.java
@@ -1,0 +1,17 @@
+package com.expediagroup.dropwizard.bundle.configuration.freemarker;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import org.apache.commons.io.IOUtils;
+
+import io.dropwizard.configuration.ConfigurationSourceProvider;
+
+public class TestConfigSourceProvider implements ConfigurationSourceProvider {
+
+    @Override
+    public InputStream open(final String config) throws IOException {
+        return IOUtils.toInputStream(config);
+    }
+
+}